+++ src/main/java/com/epam/esm/model/entity/MostlyUsedTagIdByUserId.java
@@ -1,5 +1,5 @@
 package com.epam.esm.model.entity;
 
 public interface MostlyUsedTagIdByUserId {
-    Integer getTagId();
+    Long getTagId();
 }
